Move each output formatter into its own file.
