+++ cmd/juju/status/utils.go
@@ -4,6 +4,9 @@
 package status
 
 import (
+	"fmt"
+	"reflect"
+	"sort"
 	"time"
 )
 
@@ -24,4 +27,37 @@ func formatStatusTime(t *time.Time, formatISO bool) string {
 		// Otherwise use local time.
 		return t.Local().Format("02 Jan 2006 15:04:05Z07:00")
 	}
+}
+
+// sortStringsNaturally is syntactic sugar so we can do sorts in one line.
+func sortStringsNaturally(s []string) []string {
+	sort.Sort(naturally(s))
+	return s
+}
+
+// stringKeysFromMap takes a map with keys which are strings and returns
+// only the keys.
+func stringKeysFromMap(m interface{}) (keys []string) {
+	for _, k := range reflect.ValueOf(m).MapKeys() {
+		keys = append(keys, k.String())
+	}
+	return
+}
+
+// recurseUnits calls the given recurseMap function on the given unit
+// and its subordinates (recursively defined on the given unit).
+func recurseUnits(u unitStatus, il int, recurseMap func(string, unitStatus, int)) {
+	if len(u.Subordinates) == 0 {
+		return
+	}
+	for _, uName := range sortStringsNaturally(stringKeysFromMap(u.Subordinates)) {
+		unit := u.Subordinates[uName]
+		recurseMap(uName, unit, il)
+		recurseUnits(unit, il+1, recurseMap)
+	}
+}
+
+// indent prepends a format string with the given number of spaces.
+func indent(prepend string, level int, append string) string {
+	return fmt.Sprintf("%s%*s%s", prepend, level, "", append)
 }
